Ensure output of audio is float16

+++ tali/datasets/datasets.py
@@ -367,7 +367,7 @@ class TALIMultiModalDataset(Dataset):
                 ).values()
             )[0]
 
-            frames_dict.audio = torch.Tensor(frames_dict.audio)
+            frames_dict.audio = torch.Tensor(frames_dict.audio, dtype=torch.float16)
             frames_dict.audio = frames_dict.audio.permute([1, 0])
 
             if frames_dict.audio is None:
